+++ src/components/AnimatePresence/use-presence.ts
@@ -53,6 +53,23 @@ export function usePresence(): AlwaysPresent | Present | NotPresent {
 }
 
 /**
+ * Similar to `usePresence`, except `useIsPresent` simply returns whether or not the component is present.
+ * There is no `safeToRemove` function.
+ *
+ * ```jsx
+ * import { useIsPresent } from "framer-motion"
+ *
+ * export const Component = () => {
+ *   const isPresent = useIsPresent()
+ *
+ *   useEffect(() => {
+ *     !isPresent && console.log("I've been removed!")
+ *   }, [isPresent])
+ *
+ *   return <div />
+ * }
+ * ```
+ *
  * @public
  */
 export function useIsPresent() {